requirements.txt
certifi==2024.7.4
charset-normalizer==3.3.2
customtkinter==5.2.2
darkdetect==0.8.0
idna==3.7
packaging==24.1
pillow==10.4.0
pyaes==1.6.1
Pyrogram==2.0.106
PySocks==1.7.1
python-dotenv==1.0.1
pytube==15.0.0
requests==2.32.3
TgCrypto==1.2.5
urllib3==2.2.2
